Refactor: refactoring widgets in more screen
