+++ lib/src/views/more/views/more_screen.dart
@@ -4,7 +4,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../../core/router/bottom_nav_bar.dart';
 import '../widgets/options_container_widget.dart';
 import '../widgets/user_auth_widget.dart';
-import '../widgets/user_profile_widget.dart';
+import '../widgets/user_profile/user_profile_widget.dart';
 
 class MoreScreen extends ConsumerWidget {
   const MoreScreen({super.key});
