+++ lib/src/views/more/views/settings_screen/select_language_screen.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../../../../core/enum/mxnRate.dart';
-import '../../../../../core/widgets/mxnContainer.dart';
-import '../../widgets/settings/setting_tile_widget.dart';
+import '../../widgets/settings/app_setting/select_languange_container_widget.dart';
 
 class SelectLanguageScreen extends StatelessWidget {
   const SelectLanguageScreen({super.key});
@@ -24,37 +21,14 @@ class SelectLanguageScreen extends StatelessWidget {
         centerTitle: true,
       ),
       body: SafeArea(
-          child: Center(
-        child: Column(
-          children: [
-            MxNcontainer(
-              MxN_rate: MxNRate.TWOBYONE,
-              MxN_child: Container(
-                color: Colors.white,
-                padding: EdgeInsets.symmetric(vertical: 8, horizontal: 16),
-                child: Column(
-                  children: [
-                    SettingTileWidget(
-                      title: '한국어',
-                      trailing: Icon(Icons.check),
-                    ),
-                    const Divider(),
-                    SizedBox(
-                      height: 36.w,
-                    ),
-                    Text(
-                      '향후 추가할 계획입니다',
-                      style: TextStyle(
-                        color: Colors.grey[700],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
+        child: Center(
+          child: Column(
+            children: [
+              SelectLanguangeContainerWidget(),
+            ],
+          ),
         ),
-      )),
+      ),
     );
   }
 }
