+++ lib/src/views/more/views/settings_screen/settings_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../widgets/settings/app_setting_widget.dart';
-import '../../widgets/settings/study_setting_widget.dart';
+import '../../widgets/settings/app_setting/app_setting_widget.dart';
+import '../../widgets/settings/study_setting/study_setting_widget.dart';
 
 class SettingsScreen extends StatelessWidget {
   const SettingsScreen({super.key});
