+++ lib/src/views/more/widgets/notification/notification_button.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class NotificationButton extends StatelessWidget {
+  const NotificationButton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      dense: true,
+      contentPadding: EdgeInsets.zero,
+      visualDensity: VisualDensity(vertical: -4),
+      leading: Icon(Icons.notifications_outlined),
+      title: Text(
+        '알람',
+        style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+              fontWeight: FontWeight.w500,
+            ),
+      ),
+      trailing: Icon(Icons.arrow_forward_ios),
+    );
+  }
+}
