+++ lib/src/views/more/widgets/options_container_widget.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:go_router/go_router.dart';
 
 import '../../../../core/enum/mxnRate.dart';
 import '../../../../core/widgets/mxnContainer.dart';
+import 'notification/notification_button.dart';
+import 'settings/setting_button_widget.dart';
 
 class OptionsContainerWidget extends StatelessWidget {
   const OptionsContainerWidget({super.key});
@@ -17,54 +17,12 @@ class OptionsContainerWidget extends StatelessWidget {
         padding: EdgeInsets.symmetric(vertical: 8, horizontal: 16),
         child: Column(
           children: [
-            ListTile(
-              dense: true,
-              contentPadding: EdgeInsets.zero,
-              visualDensity: VisualDensity(vertical: -4),
-              onTap: () {
-                context.go('/more/settings');
-              },
-              leading: Icon(
-                Icons.settings_outlined,
-                size: leadingIconSize,
-              ),
-              title: Text(
-                '설정',
-                style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                      fontWeight: FontWeight.w500,
-                    ),
-              ),
-              trailing: Icon(
-                Icons.arrow_forward_ios,
-                size: trailingIconSize,
-              ),
-            ),
+            SettingButtonWidget(),
             const Divider(),
-            ListTile(
-              dense: true,
-              contentPadding: EdgeInsets.zero,
-              visualDensity: VisualDensity(vertical: -4),
-              leading: Icon(
-                Icons.notifications_outlined,
-                size: leadingIconSize,
-              ),
-              title: Text(
-                '알람',
-                style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                      fontWeight: FontWeight.w500,
-                    ),
-              ),
-              trailing: Icon(
-                Icons.arrow_forward_ios,
-                size: trailingIconSize,
-              ),
-            ),
+            NotificationButton(),
           ],
         ),
       ),
     );
   }
-
-  double get trailingIconSize => 16.w;
-  double get leadingIconSize => 20.w;
 }
