+++ lib/src/views/more/widgets/user_profile/profile_image_widget.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class ProfileImageWidget extends StatelessWidget {
+  const ProfileImageWidget({
+    super.key,
+    required this.radius,
+  });
+
+  final double radius;
+
+  @override
+  Widget build(BuildContext context) {
+    return CircleAvatar(
+      radius: radius,
+      backgroundImage: AssetImage('lib/core/imgs/images/StudyDuck.png'),
+    );
+  }
+}
